// solution/frontend/src/types/trading.ts
export enum TradeSide {
  BUY = 'BUY',
  SELL = 'SELL',
}

export enum TradeType {
  MARKET = 'MARKET',
  LIMIT = 'LIMIT',
  STOP = 'STOP',
}

export enum TradeSource {
  MANUAL = 'MANUAL',
  API = 'API',
  IMPORT = 'IMPORT',
}